+++ packages/db/src/index.ts
@@ -1,5 +1,5 @@
 export * from "./model.js";
 export * from "./schema.js";
-export { db, sqlite, dbDriver } from "./connection.js";
+export { db } from "./connection.js";
 export { runMigrations } from "./migrate.js";
 export { modelFromDb, modelToDb, seedWorkspace } from "./model-io.js";